local_chatbot_web_research/backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
httpx==0.28.1
beautifulsoup4==4.13.4
lxml==6.0.0
pydantic==2.11.7
